Fix possible freeze related to custom checklistboxes (PR #9224)
* Fix possible freeze related to custom checklistboxes

`gui.nvdaControls.CustomCheckListBox` uses a custom `AccPropServer` for accessibility. However, when destroying the control, the `AccPropServer` wasn't properly unregistered. This leads to unexpected behavior, sometimes even to system freezes.

While this was fixed for `AutoWidthColumnCheckListCtrl` it was not for `CustomListBox`. Now the AccPropServer instance itself is responsible for registration and unregistration. This cleanup is triggered by binding to the the control's window destroyed wx event.