Ensure some issue template is always shown
When visiting GitHub via the `nvda/issues/new` URL no template is presented, now the bug report template is shown by default. There is a message at the beginning that explains that the `nvda/issues/new/choose` URL should be used instead. There is also URLs to jump directly to the creation of bug reports and feature requests. 
The header is not included in the HTML comment, this allows the the links to become clickable by swapping to the preview page. It will also make it easier for us to tell when someone has used this template (if they do not delete this header).