Use title for the name of the dialog.
The windows title bar includes the profile name. This gets spoken as
"NVDA Settings: (normal configuration) dialog"
Instead we want this spoken as:
"NVDA Settings dialog normal configuration"

The other benefit of the description being set is that when the dialog
receives focus, it uses the description rather than the getDialogText
implemenation in behaviours.py. In the settings dialog, more is spoken
by getDialogText than is necessary, because we have a property page
with a description which also gets spoken. This is required, so that
the explanation of the settings panels is spoken when moving from
the category list to the panel.