Add a script to start the wx inspection tool. (#9246)
* Add a script to start the wx inspection tool.

This is handy when working on the NVDA GUI.
Note: The Inspection tool has a few accessibility issues such as tab getting stuck in the details edit field. Plus that edit field is not accessible and NVDA falls back to displayModel, sometimes producing some garbled text at the end of each line.