VirtualBuffer._isNVDAObjectInApplication_noWalk: Gracefully handle rare exception from getIdentifierFromNVDAObject.
Rarely, _isNVDAObjectInApplication can walk to the desktop object.
This probably happens if the object dies while we're walking ancestors, in which case accParent will fail and NVDA will fall back to the window parent, which is the desktop.
When that happens, getIdentifierFromNVDAObject will fail.
There's not much we can do about this, so just catch the exception and return early.