MozillaCompoundTextInfo: Don't skip ignored objects when determining whether the range covers the end of any ancestors of endObj.
In order to display markers at the end of containers in braille (e.g. "lst end" at the end of lists), MozillaCompoundTextInfo needs to check the object at the end of the range (endObj) and its ancestors.
If the range is at the end of any of these objects, we mark the associated control field as being at the end of its node.
For example, if the range is at the end of a list, we mark the list's control field as being at end of node, which causes braille to display "lst end".
However, if the range is *not* at the end of one of these objects, the range also can't be at the end of any of its ancestors.
For example, if the range ends in the middle of a list item, it obviously doesn't cover the end of the list, even if the list item (as a whole) is the last item in the list.
In this case, we don't check the remaining ancestors.

Previously, this was failing where there was a single paragraph inside the last list item in a list.
The list would be marked as end of node, even if the range ended in the middle of the paragraph, resulting in a spurious "lst end" indicator before the cursor in braille.

This happened because we ignore paragraphs as irrelevant and thus don't generate control fields for them.
However, the code which checks whether the range is at the end of endObj or its ancestors only ran if there was a control field.
Because the paragraph had no control field, we'd skip it completely, not determining that the range ended in the middle of it and thus aborting the search.
We'd then check the next ancestor, and since the paragraph was the last child of the list item which was in turn the last child of the list, we'd mark all of these as end of node, including the list.

To fix this, we now run this check regardless of whether we generated a control field for an object; i.e. regardless of whether we ignored a paragraph.