OneCore voices: Use new SpeechSynthesizerOptions properties to set pitch, volume and rate lengths (PR #8934)
Fixes #7498.

Issue summary for NVDA's OneCore voices support:
- The rate setting is affected by the rate setting in Windows Speech Settings.
- The pitch range is very limited (compared with Narrator).

Previously, we used SSML in every utterance to set the base value of parameters, since there was no other way.
However, Windows 10 Fall Creators Update introduced new properties in the SpeechSynthesizerOptions class to set these parameters.

In addition to using these new properties, this commit adds rate boost to the synthesizer settings ring and added rate boost to the OneCore driver. This is disabled by default so speech should continue to be understandable. It is expected that for older versions of Windows 10 this driver should behave equally to how current master behaves.

The only case where the rate will differ from before this commit will be when someone changed the rate in the Windows 10 speech settings. More information on this in the PR #8934