Enable configuration profiles with gestures (PR #8844)
Closes #4209 

Previously configuration profiles could only be activated:
- automatically with triggers
- by manual activation in the dialog for configuration profiles

This commit adds support for activation with gestures.

1. When NVDA is started, a script is added to the class for every profile.
2. Adding a profile also adds a script
3. Removing a profile deletes it's script along with gestures that have ever been bound to it.
4. Renaming a profile removes the old script, and then adds a new one. The assigned gestures are moved from the old to the new script.

Config profiles are bound to file naming restrictions, whereas script names are bound to python variable naming rules. The latter rules are stricter. This is now taken care of by changing any thing that is not either an alphanumeric or an underscore to its base16 encoding. Doing so meets the naming requirements and avoids a name collision when having a "Test profile" and "Test_profile" profile.