Eclipse autocompletion support (PR #9297)
Autocompletion items are read from the code editor when selected.

Previously when writing code in the editor you need to focus the popup to know what items are on it. 
Detect the autocompletion popup and read the selected item as it changes.
This also includes a script to read javadoc right from the editor.