Update changes file for PR #8815
The version number for threshold is currently unknown, I have added a new section at the start of this file as a place holder for this version number.

Merges from master with modifications to this file should not be too hard, the changes from master should always come after the changes on the threshold branch. Because they are in different parts of the file there should be very few conflicts.