Don't report live regions in background tabs in Firefox.
In Firefox, all tabs have the same HWND.
Objects in background tabs do get the offscreen state, but offscreen live regions are used to report visually hidden information, so we can't filter based on that.
Therefore, if the offscreen state is set, we do an additional background check.
We get the tab document from the accessible, get the foreground tab document and compare them.
If they differ, the event is for a background tab and we thus ignore the event.