gecko_ia2 vbuf __contains__: If accChild fails, try the embedder iframe(s).
In Firefox, when an iframe document is rendered in a separate process to its embedder, accChild for accessibles in the embedded document fails.
If this is the case, we can get the embedder iframe and try accChild with the iframe's id.
There can be nested out-of-process iframes, so we keep trying this until there are no more embedders in the hierarchy (or until we hit the buffer's root id).