Create a buffer manually for writes to HID devices
By creating the buffer manually, we can ensure that data smaller than OutputReportByteLength
will be padded with null bytes. I tested ctypes marshalling by creating
a C++ dll, loading it and calling it with ctypes.

Small refactor, only hwIo.HID devices were using _writeSize. hwIo.Bulk
took the parameter, but it was only set to 0 by the hims driver. This
failed the Truth test, and len(data) was used in this case.

The refactor allows the HID complexity to stay with the HID class.
If it becomes necessary to support buffers larger than OutputReportByteLength
further refactoring will become necessary.