Python3 convert braille drivers (PR #9736)
* fix malformed newline character in brailleNote.py

Converts :
* braille driver framework.
* alva driver
* baum driver
* BrailleNote driver
* BrailliantB driver
* ecoBraille driver
* eurobraille driver
* seika driver
* superBrl driver
* freedomScientific driver
* handyTech driver
* hedoMobilLine driver
* hedoProfiLine driver
* hims driver
* lilli driver
* papenmeier driver
* papenmeier serial driver

Adds type annotations to help with fixing these issues
- braille.py 
- bdDetect.py
- hwIo.py
- brailleInput.py
- hwPortUtils.py

* Moved functions related to Input gesture in papenmeier driver. For some they depend on classes defined later in the file, others are not related to the area they are in.

* Hid device (in hwIo.py) buffers are now created manually using ctypes for writes.
By creating the buffer manually, we can ensure that data smaller than OutputReportByteLength
will be padded with null bytes. I tested ctypes marshalling by creating
a C++ dll, loading it and calling it with ctypes.

* Small refactor to hwIo.py:
 - Only class `hwIo.HID` was using `_writeSize`. class `hwIo.Bulk`
took the parameter, but only the hims driver provided it, and set it 0. This
failed the Truth test, and `len(data)` was used in this case anyway.
 - The refactor allows the HID complexity to stay with the HID class.
If it becomes necessary to support buffers larger than OutputReportByteLength
further refactoring will become necessary.

In braille display drivers, checked all usages of:
- ord()
- bytes()
- bytearray()
- str()