Initial work toward better Miranda NG support.
This commit adds a LiveTextHistoryMixin for adding a history (such as for chat applications) to LiveText. It also adds a DisplayModelLiveTextWithoutTextInfo for some controls, and begins implementation for Miranda.

It does not automatically start/stop monitoring LiveText on focusing the chat window, and the history gestures only work when the LiveText object has focus.