Python 3 one core: Don't check for sys.frozen in the check method. (PR #9857)
When running a python 2 copy from source, NVDA is unable to detect if it is running under Windows 10. Therefore, the oneCore synthesizer contains a work around in its check method, always listing the synthesizer in the synthesizer list when running from source, even on Windows versions older than Windows 10.
On Python 3 however, the manifest of the main Python targets Windows 10 and therefore, this workaround is no longer necessary.