Fix error with accPropServer
The VARIANT out-param had not been initialised correct.
Comtypes was not initialising VARIANT.vt to VT_EMPTY, and attempting to call VariantClear (oleauto.h) before assigning it the value returned from getProp.
Instead, we implement a low-level (require a this param) getProp method, which is then provided pointers to the out-params, and can initialise them correctly.