Python3: Fix save scripts in lockapp (PR #9884)
Fixes #9883

On the Windows 10 lock screen, invoked with Windows+L. The user copy of NVDA is still active. As the screen is locked, there is a set of safe scripts on the appmodule for the lock screen to make sure that the NVDA configuration isn't changed when on the lock screen. However, this SAFE_SCRIPTS set expected scripts on globalCommands.GlobalCommands to be unbound methods, so for every script, it took the __func__ attribute on the unbound method to get to the actual function. In python 3 unbound methods are gone.

SAFE_SCRIPTS now just contains the raw scripts from globalCommands.GlobalCommands, which are now functions, no longer unbound methods. When comparing a bound script with the SAFE_SCRIPTS, we're dealing with a bound method and thus we still have to traverse up to the __func__ attribute on the bound script. I added an explanatory comment for that.