Python 3: Fix error with accPropServer (PR #9900)
Fixes #9768

The VARIANT out-param was not being initialised correctly. Calling VariantClear on an uninitialised VARIANT is an error.

Comtypes was not initialising VARIANT.vt to VT_EMPTY, and attempting to call VariantClear (oleauto.h) before assigning it the value returned from getProp.
Instead, we implement a low-level getProp method by requiring an unused "this" param. This low-levl getProp method is provided pointers to the out-params, allowing it to manually initialise the VARIANT correctly before assigning a value.