Differentiate preselected from just selected text (PR #9909)
The order of selection messages for preselected and just selected text needs to be different. When a control with preselected text gains focus, we must inform them it is selected first so they do not unintentionally modify the text. When a user intentionally selects text, the 'selected' part of the message is still important (for users who may not be familiar with a command), but the text selected is more important.

This commit differentiates and documents the differences here with helper methods to make these announcements.