Python 3: Fix several local overrides of range function (PR #9912)
We were using range as a name of a local attribute or function parameter. In at least one case (uia _getBoundingRectsFromUIARange), this fails, as were also calling the range function in that method.

Looked for every instance of range as a local attribute. In most cases, rename it to textRange.