Remove .pyc and .pyo files when updating (PR #9961)
When updating NVDA, old *.pyc and *.pyo files were remaining. This is problematic for Python 2 pyc files that Python 3 tries to open. Furthermore, it causes issues such as reported in #9960, where the logHandler gets really mad if it tries to log a traceback for a file in library.zip which is also available in the installation directory.

When removing old program files, check for *.pyc and *.pyo files and get rid of them.