Ignore _expandCollapseBeforeReview in script_reviewStartOfLine. (PR #9944)
This fixes an issue in UIA consoles, invoking the "review start of line" script (`shift+numpad 1` by default) does not move the review cursor.

Expanding and collapsing the textInfo is essential for the functionality of this script (to move the textInfo to the start of the line), we should ignore the _expandCollapseBeforeReview flag.

The _expandCollapseBeforeReview flag was added in #9614. It is True everywhere except in UIA consoles.