Revert search history (PR #9993)
Revert "History in search dialog (PR #8921)"

Although useful, search history combo box does not allow languages requiring IME (input method editor) to cancel entry of search text. This affects Chinese, Japanese, Korean and other languages.

This reverts commit e3b2e797cbcdada8522bb872d55b92a265a041f0.

Keep the "Searching for text section" in the user guide as most of the section
is accurate and there are links to this section in other parts of the document.

Only remove lines that refer to search history.