Calculator: gather a list of UIA automation ID's for objects that should not allow NVDA to repeatedly announce calculation results or expressions. Re #9428.
With the advent of alwasy on top i.e. compact overlay mode in Calculator, additional UIA automation ID's were added for calculation results and expressions in overlay mode. Because of ine length concerns, the additional ID's, together with the existing ones, are now housed inside a list of known ID's causing NVDA to announce expressions repeatedly. Not only this makes it easy to add more ID's in the future (either through NVDA Core or add-ons), it improves readability for name change and notification event handler methods (and keeps the lines short).