Gecko vbuf backend: Remove old workaround for old versions of Firefox which didn't fire reorder on iframes.
In Firefox <= 3.5, a reorder event wasn't fired on an iframe when its embedded document changed (Mozilla bug 420845).
To work around this, the Gecko vbuf had code in its win event callback to retrieve the target for stateChange events and query the NODE_CHILD_OF relation to get the iframe.
This was fixed in Firefox 3.6 (10 years ago) and thus is no longer needed.
Also, this code causes us to retrieve objects unnecessarily in the win event callback, which does matter more in Firefox now that these are cross-process calls.
Although I wouldn't have thought this would have much of a performance impact, this code path does show up in Firefox profiling.