App properties: use a dedicated (private) method for fetching package info for Microsoft Store apps. Re #7894.
User32.dll::IsImmersiveProcess won't quite cut it for Store apps:
* File Explorer is seen as an immersive app but it is not really a Store application.
* Converted desktop apps are seen as not immersive but they are Store apps.
The more accurate way to determine this information is looking for app package info, which can be done by calling kernel32.dll::GetPackageFullName. This returns a string representation of app name, version, architecture, language, and app ID (the first two are used in #4259). Every Store app comes with app model XML file that provides data for this structure (Store version of NVDA is no exception). Thus use this method, which is now part of a private function in the base app module so not only product name and version setter can use it, but also Store app property can call this method.