Fix several memory leaks in the webKit vbuf backend.
When this code was originally written, we needed to hold onto the IAccessible pointers so we could map them to/from vbuf nodes.
Once WebKit implemented IAccessible2, I converted the vbuf backend to use IAccessible2 and this was no longer necessary.
Unfortunately, in the conversion, I neglected to add code to release the pointers we previously held onto.

In fixing this, I've moved everything to use CComPtr/Variant/BSTR smart pointers.