appModules/SearchUI: add support for announcing suggestion count in Windows 10 Version 1909. Re #10341.
In Windows 10 Version 1909 (build 18363) and later, File Explorer gains ability to search OneDrive content. This is powered by an embedded Windows Search feature (SearchUI.exe) with a live region announcing suggestion count. This live region has no label, but the child object does, thus coerce the name of this object to be the name of its child, which then allows live region change event to work.