Use supportedSettings instead of sizerDict to determine if options exist (PR #10504)
For a provider that only has run-time settings and uses the VisionProviderSubPanel_Wrapper, the VisionProviderSubPanel_Settings would be properly hidden if the provider is disabled. After enabling and disabling the provider, the empty panel is still shown and can receive keyboard access. Whether a panel had options was based on the sizerDict. However, for performance reasons, the sizerDict isn't cleared when updating the driver settings.

Instead use the supportedSettings property to determine if the panel has options.