Enable the Java Access Bridge for the user at startup.
Installed JREs allow users to do this from Ease of Access.
However, given that some apps now bundle their own Java VM and given that many JREs don't provide an installer (official OpenJDK is just a zip file, for example), doing this from Ease of Access isn't possible for many users.
It's also an extra hoop a user really shouldn't have to jump through.
Therefore, if this is an installed copy of NVDA and the JAB isn't already enabled for the user, we just do this quietly when initialising JABHandler.