Allow JAB messages through the Windows message filter before initialising JAB.
Internally, the JAB client dll keeps track of Java VMs.
This is done via window messages.
If a JAB client dll is initialised after a VM, it needs to learn about existing VMs.
It does this by broadcasting a message, which is picked up by Java VMs.
VMs in turn send a message back to the client DLL.
Previously, we allowed messages through the message filter *after* initialising the JAB.
That meant that messages from existing Java VMs might be rejected if they were received before the function returned.
To fix this, we allow messages through the filter before initialising the JAB.