Fix Find dialog layout (PR #10354)
Several visual issues with the Find dialog (NVDA+Ctrl+F) are fixed:
- Missing space between the "Type the text you wish to find" label and the corresponding field
- Missing space between the "Case sensitive" check box and the dialog edge
- Missing space between "OK"/"Cancel" buttons and the dialog's edge

This led to an unusual design of the dialog. It created difficulty for visually impaired people to get the expected information from the dialog. E.g. difficulty to read the text to find in the edit box because it is not properly separated from the label.