Fix checksum index in Hims braille display driver (PR #10530)
Hims driver didn't work for the Hims QBraille. When converting this driver to Python 3, a mistake was made in the _sendPacket method.

The index for the checksum was off by one. This is hard to spot, the final element in the packet list takes up two bytes. We previously used the list rather than a bytearray when overwriting the checksum index. Now we avoid converting the list twice. Because we are now overwriting the checksum in the bytearray object, and the index is one further from the end.

Co-Authored-By: Reef Turner <feerrenrut@users.noreply.github.com>